+++ objects/GameObjectManager.h
@@ -11,7 +11,7 @@ public:
   GameObjectManager();
   ~GameObjectManager();
 
-  void update(double dt);
+  void update(float dt);
 
   void addObject(GameObject* object);
   void removeObject(GameObject* object);
@@ -23,12 +23,12 @@ private:
   std::vector<GameObject*> objects;
   sf::RenderWindow* window;
 
-  const double fixedUpdateDt;
-  double timeAccumulator;
+  const float fixedUpdateDt;
+  float timeAccumulator;
   int ticks;
 
-  void variableUpdate(double dt);
-  void fixedUpdate(double dt);
+  void variableUpdate(float dt);
+  void fixedUpdate(float dt);
 
   void checkCollisions();
 };
