+++ objects/components/BoxCollider.h
@@ -9,13 +9,13 @@ class GameObject;
 class Transform;
 
 struct BoundingRectangle {
-  double left, right, top, bottom;
+  float left, right, top, bottom;
 
-  double width() const {
+  float width() const {
     return right - left;
   }
 
-  double height() const {
+  float height() const {
     return bottom - top;
   }
 };
@@ -27,16 +27,16 @@ public:
 
   bool collidesWith(BoundingRectangle other);
 
-  Vec2<double> getPenetrationVector(const std::vector<GameObject*>& objects);
+  Vec2<float> getPenetrationVector(const std::vector<GameObject*>& objects);
 
-  static BoundingRectangle getBoundingRectangle(const std::vector<Vec2<double>>& mesh);
+  static BoundingRectangle getBoundingRectangle(const std::vector<Vec2<float>>& mesh);
 
 private:
   Transform* transform;
 
-  Vec2<double> getPenetration(GameObject* object);
-  static Vec2<double> getTheoreticalPenetration(BoundingRectangle boundingRectangle, GameObject* object);
-  static Vec2<double> getActualPenetration(BoundingRectangle r1, BoundingRectangle r2);
+  Vec2<float> getPenetration(GameObject* object);
+  static Vec2<float> getTheoreticalPenetration(BoundingRectangle boundingRectangle, GameObject* object);
+  static Vec2<float> getActualPenetration(BoundingRectangle r1, BoundingRectangle r2);
 };
 
 #endif //PLATFORMER_BOXCOLLIDER_H
